python_service/requirements.txt
absl-py==2.4.0
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deepface==0.0.98
filelock==3.24.3
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.78.1
gunicorn==25.1.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
mysql-connector-python==9.6.0
namex==0.1.0
numpy==2.4.2
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
protobuf==6.33.5
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
retina-face==0.0.17
rich==14.3.3
setuptools==82.0.0
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
